replaced error handlings with custom error classes

+++ quesadiya/commands/delete.py
@@ -1,5 +1,9 @@
 import click
 
+from quesadiya.errors import ProjectNotExistError
+from quesadiya.errors import ProjectRunningError
+from quesadiya.errors import AuthenticationError
+
 from quesadiya.db import factory
 from quesadiya import utils
 import quesadiya
@@ -9,22 +13,18 @@ import os
 import sys
 
 
-# TODO: add code to check if a project is running
-# TODO: add function to print common error message
 def operator(project_name):
     admin_interface = factory.get_admindb_interface()
     if not admin_interface.check_project_exists(project_name):
-        click.echo(
-            "Project ({}) doesn't exist. Check all project names by "
-            "'quesadiya inspect all'".format(project_name)
-        )
-        return
+        raise ProjectNotExistError(project_name)
+    if admin_interface.is_project_running(project_name):
+        raise ProjectRunningError(project_name, "`quesadiya delete`")
     if not utils.admin_auth(admin_interface, project_name):
-        sys.exit("Invalid name or password for {}".format(project_name))
+        raise AuthenticationError(project_name)
     conf = click.confirm("Are you sure you want to continue?")
     if conf:
         admin_interface.delete_project(project_name)
-        click.echo('Successfully deleted {}'.format(project_name))
+        click.echo("Successfully deleted {}".format(project_name))
         shutil.rmtree(
             os.path.join(quesadiya.get_projects_path(), project_name)
         )
